Add resource2header Perl script
This change introduces the resource2header script. This script can be
used to convert an arbitrary binary file into a C/C++ header file
containing a variable with this file's data and another one representing
its size.
This header can be used to embedd arbitrary binary data statically
into a program, which is helpful if no external dependencies are to be
shipped or if opening external files is not an option, for example.